Remove width_ and height_ from various VideoDecoder implementations.
Updated unit tests as well.

BUG=none
TEST=media_unittests

Review URL: http://codereview.chromium.org/6624068

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@77336 0039d316-1c4b-4281-b951-d872f2087c98

+++ media/base/limits.h
@@ -1,5 +1,5 @@
 FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
-// Copyright (c) 2010 The Chromium Authors. All rights reserved.
+// Copyright (c) 2011 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -14,8 +14,8 @@ namespace media {
 
 struct Limits {
   // For video.
-  static const size_t kMaxDimension = (1 << 15) - 1;  // 32767
-  static const size_t kMaxCanvas = (1 << (14 * 2));  // 16384 x 16384
+  static const int kMaxDimension = (1 << 15) - 1;  // 32767
+  static const int kMaxCanvas = (1 << (14 * 2));  // 16384 x 16384
 
   // Total number of video frames which are populating in the pipeline.
   static const size_t kMaxVideoFrames = 4;
